+++ __mocks__/data.ts
@@ -12,7 +12,7 @@ export const dummyDomain = {
    lastUpdated: '2022-11-11T10:00:32.243',
    added: '2022-11-11T10:00:32.244',
    tags: '',
-   scrape_enabled: true,
+   scrapeEnabled: true,
    notification: true,
    notification_interval: 'daily',
    notification_emails: '',