+++ database/models/domain.ts
@@ -28,7 +28,7 @@ class Domain extends Model {
    tags!: string;
 
    @Column({ type: DataType.BOOLEAN, allowNull: false, defaultValue: true })
-   scrape_enabled!: boolean;
+   scrapeEnabled!: boolean;
 
    @Column({ type: DataType.BOOLEAN, allowNull: true, defaultValue: true })
    notification!: boolean;