+++ pages/api/notify.ts
@@ -63,7 +63,7 @@ const notify = async (req: NextApiRequest, res: NextApiResponse<NotifyResponse>)
          const theDomain = await Domain.findOne({ where: { domain: reqDomain } });
          if (theDomain) {
             const domainPlain = theDomain.get({ plain: true }) as DomainType;
-            if (domainPlain.scrape_enabled !== false && domainPlain.notification !== false) {
+            if (domainPlain.scrapeEnabled !== false && domainPlain.notification !== false) {
                await sendNotificationEmail(domainPlain, normalizedSettings);
             }
          }
@@ -72,7 +72,7 @@ const notify = async (req: NextApiRequest, res: NextApiResponse<NotifyResponse>)
          if (allDomains && allDomains.length > 0) {
             const domains = allDomains.map((el) => el.get({ plain: true }));
             for (const domain of domains) {
-               if (domain.scrape_enabled !== false && domain.notification !== false) {
+               if (domain.scrapeEnabled !== false && domain.notification !== false) {
                   await sendNotificationEmail(domain, normalizedSettings);
                }
             }